Remove date code
I am using jekyll's blog-aware functions to feed markdown pages and get out nicely formatted content. The thing is, this is a static webpage, not a blog, so the date codes are false. They should not be visible on the "posts" because I am only using them so that I can write my pages in markdown.