added welcome text to index page
THis is a really ugly hack. I don't know how to make the index page automatically redirect to the home page blog entry.  So I did a bad thing. I copied the home blog entry text and pasted it directly into the index page.  Now the home text lives in two places: in the home.md file, and on the index page.

The gods of programming frown upon my sin today.